✨ feat(extensions): convert answer to directory layout with tests
